Items now have an ID
